+++ gr-blocks/lib/tagged_stream_to_pdu_impl.cc
@@ -49,7 +49,7 @@ namespace gr {
     }
 
     int
-    tagged_stream_to_pdu_impl::work (int noutput_items,
+    tagged_stream_to_pdu_impl::work (size_t noutput_items,
                        gr_vector_int &ninput_items,
                        gr_vector_const_void_star &input_items,
                        gr_vector_void_star &output_items)
