+++ gr-blocks/lib/tsb_vector_sink_impl.cc
@@ -68,7 +68,7 @@ namespace gr {
 
     template <class T>
     int
-    tsb_vector_sink_impl<T> ::work (int noutput_items,
+    tsb_vector_sink_impl<T> ::work (size_t noutput_items,
                        gr_vector_int &ninput_items,
                        gr_vector_const_void_star &input_items,
                        gr_vector_void_star &output_items)
