+++ gr-blocks/lib/tsb_vector_sink_impl.h
@@ -45,7 +45,7 @@ template<class T>
       std::vector<std::vector<T> > data() const;
       std::vector<tag_t> tags() const;
 
-      int work(int noutput_items,
+      int work(size_t noutput_items,
 		       gr_vector_int &ninput_items,
 		       gr_vector_const_void_star &input_items,
 		       gr_vector_void_star &output_items);
