+++ gr-digital/lib/ofdm_carrier_allocator_cvc_impl.cc
@@ -147,7 +147,7 @@ namespace gr {
     }
 
     int
-    ofdm_carrier_allocator_cvc_impl::work (int noutput_items,
+    ofdm_carrier_allocator_cvc_impl::work (size_t noutput_items,
                        gr_vector_int &ninput_items,
                        gr_vector_const_void_star &input_items,
                        gr_vector_void_star &output_items)
