+++ gr-digital/lib/ofdm_carrier_allocator_cvc_impl.h
@@ -64,10 +64,10 @@ namespace gr {
       const int fft_len() { return d_fft_len; };
       std::vector<std::vector<int> > occupied_carriers() { return d_occupied_carriers; };
 
-      int work(int noutput_items,
-		 gr_vector_int &ninput_items,
-		 gr_vector_const_void_star &input_items,
-		 gr_vector_void_star &output_items);
+      int work(size_t noutput_items,
+         gr_vector_int &ninput_items,
+         gr_vector_const_void_star &input_items,
+         gr_vector_void_star &output_items);
     };
 
   } // namespace digital
