+++ gr-digital/lib/ofdm_cyclic_prefixer_impl.h
@@ -51,10 +51,10 @@ namespace gr {
       ofdm_cyclic_prefixer_impl(size_t input_size, size_t output_size, int rolloff_len, const std::string &len_tag_key);
       ~ofdm_cyclic_prefixer_impl();
 
-      int work(int noutput_items,
-		       gr_vector_int &ninput_items,
-		       gr_vector_const_void_star &input_items,
-		       gr_vector_void_star &output_items);
+      int work(size_t noutput_items,
+               gr_vector_int &ninput_items,
+               gr_vector_const_void_star &input_items,
+               gr_vector_void_star &output_items);
     };
 
   } // namespace digital
