+++ gr-digital/lib/ofdm_serializer_vcc_impl.cc
@@ -144,7 +144,7 @@ namespace gr {
     }
 
     int
-    ofdm_serializer_vcc_impl::work (int noutput_items,
+    ofdm_serializer_vcc_impl::work (size_t noutput_items,
                        gr_vector_int &ninput_items,
                        gr_vector_const_void_star &input_items,
                        gr_vector_void_star &output_items)
