+++ gr-digital/lib/ofdm_serializer_vcc_impl.h
@@ -59,10 +59,10 @@ namespace gr {
       );
       ~ofdm_serializer_vcc_impl();
 
-      int work(int noutput_items,
-		       gr_vector_int &ninput_items,
-		       gr_vector_const_void_star &input_items,
-		       gr_vector_void_star &output_items);
+      int work(size_t noutput_items,
+               gr_vector_int &ninput_items,
+               gr_vector_const_void_star &input_items,
+               gr_vector_void_star &output_items);
     };
 
   } // namespace digital
