+++ gr-digital/lib/protocol_formatter_bb_impl.cc
@@ -67,7 +67,7 @@ namespace gr {
     }
 
     int
-    protocol_formatter_bb_impl::work(int noutput_items,
+    protocol_formatter_bb_impl::work(size_t noutput_items,
                                      gr_vector_int &ninput_items,
                                      gr_vector_const_void_star &input_items,
                                      gr_vector_void_star &output_items)
