+++ gr-digital/lib/protocol_formatter_bb_impl.h
@@ -44,7 +44,7 @@ namespace gr {
       int calculate_output_stream_length(const gr_vector_int &ninput_items) {
         return d_format->header_nbytes(); };
 
-      int work(int noutput_items,
+      int work(size_t noutput_items,
                gr_vector_int &ninput_items,
                gr_vector_const_void_star &input_items,
                gr_vector_void_star &output_items);
