+++ gr-fec/lib/tagged_encoder_impl.cc
@@ -76,7 +76,7 @@ namespace gr {
     }
 
     int
-    tagged_encoder_impl::work(int noutput_items,
+    tagged_encoder_impl::work(size_t noutput_items,
                               gr_vector_int& ninput_items,
                               gr_vector_const_void_star &input_items,
                               gr_vector_void_star &output_items)
